requirements.txt
Flask
Flask-Dance
opencv-python
PyYAML